requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
apscheduler>=3.10.0
httpx>=0.24.0
python-dotenv>=1.0.0
websockets>=12.0
